* subversion/svnserve/cyrus_auth.c   (canonicalize_username): Use the right type to avoid a size mismatch error.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1349716 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/svnserve/cyrus_auth.c
@@ -61,7 +61,7 @@ static int canonicalize_username(sasl_conn_t *conn,
                                  char *out, /* the output buffer */
                                  unsigned out_max, unsigned *out_len)
 {
-  int realm_len = strlen(user_realm);
+  size_t realm_len = strlen(user_realm);
   char *pos;
 
   *out_len = inlen;
